feat: tx hash
